Moved the examples directory to the root
